Exclude a duplicate Provides symbol
The `perl(HTTP::Message::Common)` symbol is already provided by
`perl-HTTP-Message`, and that package should be the only provider.

RPM's `perl.prov` generator correctly detects the duplicate when scanning
`Apache-Test/lib/Apache/TestRequest.pm` installed by this package, but after
examining the source file, I've determined that it's not intended to be a
symbol to Provide.

The easiest fix is to exclude the Provide via the spec file directly.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>